pyproject.toml
[project]
name = "tessan-medical-chatbot"
version = "0.1.0"
description = "Medical chatbot for Tessan"
authors = [
    {name = "YOUSSEF"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-genai (>=1.60.0,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "langchain (>=1.2.7,<2.0.0)",
    "langchain-core (>=1.2.7,<2.0.0)",
    "langchain-google-genai (>=4.2.0,<5.0.0)",
    "streamlit (>=1.53.1,<2.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
